requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
deprecation==2.1.0
discord.py==2.4.0
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.0
frozenlist==1.4.1
fsspec==2024.9.0
gotrue==2.8.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
hyperframe==6.0.1
idna==3.8
jiter==0.5.0
multidict==6.0.5
openai==1.44.0
packaging==24.1
postgrest==0.16.11
pydantic==2.9.0
pydantic_core==2.23.2
pydotenv==0.0.7
python-dateutil==2.9.0.post0
PyYAML==6.0.2
realtime==2.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websockets==12.0
yarl==1.10.0
